derived productInPlanController and Service out of DietPlanController and service
