+++ VirtusFitWeb/Views/DietPlan/DailyProductList.cshtml
@@ -49,11 +49,11 @@
                         @Html.DisplayFor(modelItem => item.TotalCalories)
                     </td>
                     <td>
-                        <a class="btn btn-info btn-block" asp-controller="DietPlan" asp-action="ProductDetails" asp-route-id=@Model.DietPlanId
+                        <a class="btn btn-info btn-block" asp-controller="ProductInPlan" asp-action="ProductDetails" asp-route-id=@Model.DietPlanId
                            asp-route-dayNumber="@Model.DayNumber" asp-route-ordinalNumber="@item.OrdinalNumber">Details</a>
-                        <a class="btn btn-danger btn-block" asp-controller="DietPlan" asp-action="DeleteProduct" asp-route-id=@Model.DietPlanId
+                        <a class="btn btn-danger btn-block" asp-controller="ProductInPlan" asp-action="DeleteProduct" asp-route-id=@Model.DietPlanId
                            asp-route-dayNumber="@Model.DayNumber" asp-route-ordinalNumber="@item.OrdinalNumber">Delete</a>
-                        <a class="btn btn-primary btn-block" asp-controller="DietPlan" asp-action="EditProductInPlan" asp-route-id=@Model.DietPlanId
+                        <a class="btn btn-primary btn-block" asp-controller="ProductInPlan" asp-action="EditProductInPlan" asp-route-id=@Model.DietPlanId
                            asp-route-dayNumber="@Model.DayNumber" asp-route-ordinalNumber="@item.OrdinalNumber">Edit</a>
                     </td>
                 </tr>
@@ -96,7 +96,7 @@ else
     </table>
 }
 <p>
-    <a asp-action="ProductsToAdd" asp-route-id="@Model.DietPlanId" asp-route-dayNumber="@Model.DayNumber" class="btn btn-primary btn-block col-sm-2">Add product</a>
+    <a asp-controller="ProductInPlan" asp-action="ProductsToAdd" asp-route-id="@Model.DietPlanId" asp-route-dayNumber="@Model.DayNumber" class="btn btn-primary btn-block col-sm-2">Add product</a>
     <a asp-action="DayList" asp-route-id="@Model.DietPlanId" class="btn btn-info btn-block col-sm-2">Back to day list</a>
 
 </p>
