+++ VirtusFitWeb/Views/ProductInPlan/DeleteProduct.cshtml
@@ -1,12 +1,13 @@
 ﻿@model ProductInPlanViewModel
 
 @{
-    ViewData["Title"] = "Product details";
+    ViewData["Title"] = "DeletePlan product";
 }
 
-<h1>Product details</h1>
+<h1>DeletePlan product</h1>
 
 <h4>Product &#8470; @Model.ProductInPlan.OrdinalNumber | Diet Plan &#8470; @Model.DietPlanId | Day &#8470; @Model.DayNumber (@Model.Date)</h4>
+<h4>Are You sure You want to delete this?</h4>
 <hr />
 <h6>Nutrients values refer to 100ml/100mg portions</h6>
 <div class="row">
@@ -104,12 +105,12 @@
 
         </div>
         <div class="form-group">
-            <a asp-controller="DietPlan" asp-action="EditProductInPlan" asp-route-id="@Model.DietPlanId" asp-route-dayNumber="@Model.DayNumber"
-               asp-route-ordinalNumber="@Model.ProductInPlan.OrdinalNumber" class="btn btn-primary btn-block">Edit</a>
-            <a asp-controller="DietPlan" asp-action="DeleteProduct" asp-route-id="@Model.DietPlanId" asp-route-dayNumber="@Model.DayNumber"
-               asp-route-ordinalNumber="@Model.ProductInPlan.OrdinalNumber" class="btn btn-danger btn-block">Delete</a>
-            <a asp-controller="DietPlan" asp-action="DailyProductList" asp-route-id="@Model.DietPlanId" asp-route-dayNumber="@Model.DayNumber"
-               class="btn btn-info btn-block">Back to product list</a>
+            <form asp-action="DeleteProduct" asp-controller="ProductInPlan" asp-route-id="@Model.DietPlanId" asp-route-dayNumber="@Model.DayNumber" 
+                  asp-route-ordinalNumber="@Model.ProductInPlan.OrdinalNumber" method="post">
+                <input type="submit" value="Delete" class="btn btn-danger btn-block" />
+                <a asp-controller="DietPlan" asp-action="DailyProductList" asp-route-id="@Model.DietPlanId" asp-route-dayNumber="@Model.DayNumber"
+                   class="btn btn-info btn-block">Back to product list</a>
+            </form>
         </div>
 
     </div>
