+++ VirtusFitWeb/Views/ProductInPlan/EditProductInPlan.cshtml
@@ -1,16 +1,15 @@
 ﻿@model ProductInPlanViewModel
 
 @{
-    ViewData["Title"] = "Add product to plan";
+    ViewData["Title"] = "Edit product in plan";
 }
 
 @{
-    var productId = Model.ProductInPlan.Product.ProductId;
+    var currentProductOrdinalNumber = Model.ProductInPlan.OrdinalNumber;
 }
-<h1>Add product to plan</h1>
+<h1>Edit product in daily diet plan</h1>
 
-<h4>You are adding following product to Diet Plan &#8470; @Model.DietPlanId | Day &#8470; @Model.DayNumber (@Model.Date)</h4>
-
+<h4>Product &#8470; @Model.ProductInPlan.OrdinalNumber | Diet Plan &#8470; @Model.DietPlanId | Day &#8470; @Model.DayNumber (@Model.Date)</h4>
 <hr />
 <h6>Nutrients values refer to 100ml/100mg portions</h6>
 <div class="row">
@@ -98,7 +97,8 @@
                 </tr>
             </tbody>
         </table>
-        <form asp-controller="DietPlan" asp-action="AddProductToPlan" asp-route-id="@Model.DietPlanId" asp-route-dayNumber="@Model.DayNumber" asp-route-productId="@productId" method="post">
+        <form asp-antiforgery="true" asp-controller="ProductInPlan" asp-action="EditProductInPlan" asp-route-id="@Model.DietPlanId" asp-route-dayNumber="@Model.DayNumber" 
+              asp-route-currentProductOrdinalNumber="@currentProductOrdinalNumber" method="post">
             <div asp-validation-summary="ModelOnly" class="text-danger"></div>
             <div class="form-group">
                 <label asp-for="ProductInPlan.PortionSize" class="control-label"></label>
@@ -111,9 +111,9 @@
                 <span asp-validation-for="ProductInPlan.NumberOfPortions" class="text-danger"></span>
             </div>
             <div class="form-group">
-                <input type="submit" value="Add" class="btn btn-primary btn-block" />
-                <a asp-controller="DietPlan" asp-action="ProductsToAdd" asp-route-id="@Model.DietPlanId" asp-route-dayNumber="@Model.DayNumber" 
-                   asp-route-productId="@productId" class="btn btn-info btn-block">Back to product list</a>
+                <input type="submit" value="Confirm" class="btn btn-primary btn-block" />
+                <a asp-controller="DietPlan" asp-action="DailyProductList" asp-route-id="@Model.DietPlanId" asp-route-dayNumber="@Model.DayNumber"
+                   class="btn btn-info btn-block">Back to product list</a>
             </div>
         </form>
     </div>
