+++ VirtusFitWeb/Views/ProductInPlan/ProductsToAdd.cshtml
@@ -94,7 +94,7 @@
             </td>
             <td>
 
-                <a asp-controller="DietPlan" asp-action="AddProductToPlan" asp-route-productId=@item.ProductId
+                <a asp-controller="ProductInPlan" asp-action="AddProductToPlan" asp-route-productId=@item.ProductId
                    asp-route-id=@Model.DietPlanId asp-route-dayNumber=@Model.DayNumber class="btn btn-primary btn-block">Add</a>
 
             </td>
